+++ Tabby/src/Config.h
@@ -4,3 +4,4 @@
 #define MAX_SOUNDS 32
 #define MAX_GAMEPADS 4
 #define MAX_GAMEPAD_AXIS 8
+#define FIXED_UPDATE_RATE 60 // This is the rate of FixedUpdate for scripts.