+++ Tabby/src/Tabby/Scene/ScriptableEntity.h
@@ -23,7 +23,7 @@ protected:
     virtual void OnDestroy() { }
     virtual void Update(Timestep ts) { }
     virtual void LateUpdate(Timestep ts) { }
-    // virtual void FixedUpdate(Timestep ts) { }
+    virtual void FixedUpdate(Timestep ts) { }
 
     // Physics Callback
     virtual void OnCollisionEnter(ContactCallback contact) { }